Add homepage, back to homepage button; move workout list to homepage
