+++ FitAppMVVM/Presentation/HomePage.xaml
@@ -1,0 +1,33 @@
+﻿<Page
+    x:Class="FitAppMVVM.Presentation.HomePage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:FitAppMVVM.Presentation"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <ListView ItemsSource="{Binding Workouts}">
+            <ListView.ItemTemplate>
+                <DataTemplate>
+                    <StackPanel Margin="20,20,0,0">
+                        <TextBlock Text="{Binding Name}" />
+                        <TextBlock Text="{Binding Notes}" />
+                    </StackPanel>
+                </DataTemplate>
+            </ListView.ItemTemplate>
+        </ListView>
+        <StackPanel
+            Grid.Row="1"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            Spacing="16">
+
+
+            <Button Content="Add New Workout" Click="GoToWorkoutsPage_Click" />
+            
+        </StackPanel>
+    </Grid>
+</Page>
