+++ FitAppMVVM/Presentation/HomePage.xaml.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices.WindowsRuntime;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Controls.Primitives;
+using Microsoft.UI.Xaml.Data;
+using Microsoft.UI.Xaml.Input;
+using Microsoft.UI.Xaml.Media;
+using Microsoft.UI.Xaml.Navigation;
+using Windows.Foundation;
+using Windows.Foundation.Collections;
+
+// The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
+
+namespace FitAppMVVM.Presentation;
+/// <summary>
+/// An empty page that can be used on its own or navigated to within a Frame.
+/// </summary>
+public sealed partial class HomePage : Page
+{
+    public HomePage()
+    {
+        this.InitializeComponent();
+        this.DataContext = new HomePageViewModel();
+    }
+
+    private void GoToWorkoutsPage_Click(object sender, RoutedEventArgs e)
+    {
+        // Navigate to WorkoutPage
+        this.Frame.Navigate(typeof(WorkoutPage));
+    }
+}
