+++ FitAppMVVM/Presentation/WorkoutPage.xaml
@@ -9,16 +9,7 @@
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
     <Grid>
-        <ListView ItemsSource="{Binding Workouts}">
-            <ListView.ItemTemplate>
-                <DataTemplate>
-                    <StackPanel>
-                        <TextBlock Text="{Binding Name}" />
-                        <TextBlock Text="{Binding Notes}" />
-                    </StackPanel>
-                </DataTemplate>
-            </ListView.ItemTemplate>
-        </ListView>
+        
 
 
         <StackPanel Grid.Row="1"
@@ -29,7 +20,13 @@
 
             <TextBox Text="{Binding Name, Mode=TwoWay}" PlaceholderText="Enter the name:" />
             <TextBox Text="{Binding Notes, Mode=TwoWay}" PlaceholderText="Enter a note:" />
-            <Button Content="Add A Workout" Command="{Binding AddWorkoutCommand}" />
+
+            <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Spacing="10">
+                <Button Content="Back To Home Page" Click="GoToHomePage_Click" />
+                <Button Content="Add Workout" Command="{Binding AddWorkoutCommand}" />
+            </StackPanel>
+            
+            
             
         </StackPanel>
